+++ src/generator.js
@@ -1,12 +1,14 @@
 const vscode = require('vscode');
 const fs = require('fs');
+
 const scopes = require('../data/scopes.json');
+const pkg = require('../data/generatedThemePackage.json');
 
 
 const writeThemetoFile = (theme) => {
     fs.writeFile(`${process.env['HOME']}/.vscode/extensions/generatedTheme/themes/generated-color-theme.json` , JSON.stringify(theme, null, 4), (err) => {
         if(err) {
-            vscode.window.showErrorMessage('Failed to write theme file.')
+            vscode.window.showErrorMessage('Failed to write theme file.');
             return console.log(err.message);
         }
         console.log(`Custom Theme Generated!`);
@@ -21,7 +23,7 @@ const generateColorTheme = (colors) => {
     }
 
     // Grab primary and foreground for defaults
-    const foreground = colors.hasOwnProperty('foreground') ? colors.foreground : "#fff";
+    const foreground = colors.hasOwnProperty('foreground') ? colors.foreground : "#ffffff";
     const primary = colors.hasOwnProperty('primary') ? colors.primary : foreground;
     const background = colors.hasOwnProperty('background') ? colors.background : "#212121";
     const backgroundSecondary = colors.hasOwnProperty('backgroundSecondary') ? colors.backgroundSecondary : background;
@@ -60,9 +62,34 @@ const generateColorTheme = (colors) => {
     return theme;
 }
 
+const directorySetup = () => {
+    if (!fs.existsSync(`${process.env['HOME']}/.vscode/extensions/generatedTheme/themes`)) {
+        console.log('not exists' + `${process.env['HOME']}/.vscode/extensions/generatedTheme/themes`);
+        fs.mkdir(`${process.env['HOME']}/.vscode/extensions/generatedTheme`, (err) => {
+            if (err) {
+                vscode.window.showErrorMessage(`Failed to properly create necessary file structure. VSThemeGenerator Requires ${process.env['HOME']}/.vscode/extensions/generatedTheme/themes to exist.`)
+                return console.log(err.message);
+            }
+            fs.writeFile(`${process.env['HOME']}/.vscode/extensions/generatedTheme/package.json`, JSON.stringify(pkg, null, 4), (err) => {
+                if (err) {
+                    vscode.window.showErrorMessage(`Failed to properly create necessary file structure. VSThemeGenerator Requires ${process.env['HOME']}/.vscode/extensions/generatedTheme/themes to exist.`)
+                    return console.log(err.message);
+                }
+            });
+            fs.mkdir(`${process.env['HOME']}/.vscode/extensions/generatedTheme/themes`, (err) => {
+                if (err) {
+                    vscode.window.showErrorMessage(`Failed to properly create necessary file structure. VSThemeGenerator Requires ${process.env['HOME']}/.vscode/extensions/generatedTheme/themes to exist.`)
+                    return console.log(err.message);
+                }
+            });
+        });
+    }
+}
+
 module.exports = {
     generateColorTheme,
-    writeThemetoFile
+    writeThemetoFile,
+    directorySetup
 }