+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 All notable changes to the "vsthemegenerator" extension will be documented in this file.
 
+## 1.1.0
+- Fixing bugs causing error to be displayed when file structure existed
+- Adding proper color validation on the strings used to create the theme
+- Added `Random Theme` command to generate random color themes (*Warning: they may be CrAzY*)
+- Adding Ruby, Typescript, Markdown, and HTML language support
+
 ## 1.0.0
 - Initial release